Fix when empty content is provided, and fix updated behavior
When you simply want to create an empty hint (e.g. for EC2) and the hint
doesn't yet exist with non-empty content, it will report as up to date
because of a bad check on the content.

The content check could've been fixed, but I'd rather rely on the
resources just doing their thing instead. This simplifies the LWRP
considerably and also improves the reporting of whether anything was
actually done.