Properly handle hints where content isn't specified
Previously we'd get a deprecation warning from Chef and the content
wouldn't ever get created due to the existing resource comparison
method. By changing the default value to nil and parsing an empty JSON
file to {} we avoid the comparison issue. We also now avoid passing
content of nil to the file resource by interpretting nil to really mean
content of ''